pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "identity-access-service"
version = "0.1.0"
description = "Autenticação, autorização e auditoria para APIs internas"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "gunicorn>=23.0.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.0.0",
  "alembic>=1.14.0",
  "bcrypt>=4.0.0",
  "email-validator>=2.0.0",
  "pyjwt>=2.8.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
only-include = ["src/identity_access_service"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
